ConsumerBase2.cs 에 Pay(주문에 맞는 가격 지불)함수 추가

+++ Assets/Scripts/Data/Player.cs
@@ -6,12 +6,12 @@ public class Player
 {
     public string name;
     public int repute;
-    public int money;
+    public float money;
     //public List<Ingredient> inventory;
     public int customTruck;
     //장사일지도 추가되어야함..
 
-    public Player(string name, int repute, int money, int customTruck)
+    public Player(string name, int repute, float money, int customTruck)
     {
         this.name = name;
         this.repute = repute;
